requirements.txt
flask
pandas
numpy
scikit-learn
opencv-python
RPi.GPIO
